+++ library/dellemc/idrac/dellemc_idrac_lc_attributes.py
@@ -3,8 +3,8 @@
 
 #
 # Dell EMC OpenManage Ansible Modules
-# Version 2.1.5
-# Copyright (C) 2018-2020 Dell Inc. or its subsidiaries. All Rights Reserved.
+# Version 3.0.0
+# Copyright (C) 2018-2021 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 #
@@ -20,7 +20,7 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
 DOCUMENTATION = """
 ---
 module: dellemc_idrac_lc_attributes
-short_description: Enable or disable Collect System Inventory on Restart (CSIOR) property for all iDRAC/LC jobs.
+short_description: Enable or disable Collect System Inventory on Restart (CSIOR) property for all iDRAC/LC jobs
 version_added: "2.3.0"
 description:
     -  This module is responsible for enabling or disabling of Collect System Inventory on Restart (CSIOR)
@@ -70,28 +70,46 @@ requirements:
     - "omsdk"
     - "python >= 2.7.5"
 author: "Felix Stephen (@felixs88)"
-
+notes:
+    - Run this module from a system that has direct access to DellEMC iDRAC.
+    - This module supports C(check_mode).
 """
 
 EXAMPLES = """
 ---
 - name: Set up iDRAC LC Attributes
   dellemc_idrac_lc_attributes:
-       idrac_ip:   "xx.xx.xx.xx"
-       idrac_user: "xxxx"
-       idrac_password:  "xxxxxxxx"
-       share_name: "xx.xx.xx.xx:/share"
-       share_password:  "xxxxxxxx"
-       share_user: "xxxx"
+       idrac_ip:   "192.168.0.1"
+       idrac_user: "user_name"
+       idrac_password:  "user_password"
+       share_name: "192.168.0.1:/share"
        share_mnt: "/mnt/share"
-       csior: "xxxxxxx"
+       csior: "Enabled"
 """
 
 RETURNS = """
-dest:
+msg:
     description: Collect System Inventory on Restart (CSIOR) property for all iDRAC/LC jobs is configured.
-    returned: success
-    type: string
+    returned: always
+    type: dict
+    sample: {
+        "CompletionTime": "2020-03-30T00:06:53",
+        "Description": "Job Instance",
+        "EndTime": null,
+        "Id": "JID_1234512345",
+        "JobState": "Completed",
+        "JobType": "ImportConfiguration",
+        "Message": "Successfully imported and applied Server Configuration Profile.",
+        "MessageArgs": [],
+        "MessageArgs@odata.count": 0,
+        "MessageId": "SYS053",
+        "Name": "Import Configuration",
+        "PercentComplete": 100,
+        "StartTime": "TIME_NOW",
+        "Status": "Success",
+        "TargetSettingsURI": null,
+        "retval": true
+    }
 """