+++ library/dellemc/idrac/idrac_timezone_ntp.py
@@ -3,8 +3,8 @@
 
 #
 # Dell EMC OpenManage Ansible Modules
-# Version 2.1.1
-# Copyright (C) 2018-2020 Dell Inc. or its subsidiaries. All Rights Reserved.
+# Version 3.0.0
+# Copyright (C) 2018-2021 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 #
@@ -20,10 +20,10 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
 DOCUMENTATION = """
 ---
 module: idrac_timezone_ntp
-short_description: Configures time zone and NTP on iDRAC.
+short_description: Configures time zone and NTP on iDRAC
 version_added: "2.9"
 description:
-    - This module is responsible for configuring the iDRAC timezone attributes.
+    - This module allows to configure time zone and NTP on iDRAC.
 options:
     idrac_ip:
         required: True
@@ -90,6 +90,9 @@ requirements:
 author:
     - "Felix Stephen (@felixs88)"
     - "Anooja Vardhineni (@anooja-vardhineni)"
+notes:
+    - Run this module from a system that has direct access to DellEMC iDRAC.
+    - This module supports C(check_mode).
 """
 
 EXAMPLES = """