+++ library/dellemc/ome/ome_firmware_baseline.py
@@ -3,8 +3,8 @@
 
 #
 # Dell EMC OpenManage Ansible Modules
-# Version 2.1.5
-# Copyright (C) 2019-2020 Dell Inc. or its subsidiaries. All Rights Reserved.
+# Version 3.0.0
+# Copyright (C) 2019-2021 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 #
@@ -20,7 +20,7 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
 DOCUMENTATION = r'''
 ---
 module: ome_firmware_baseline
-short_description: Create a firmware baseline on OpenManage Enterprise.
+short_description: Create a firmware baseline on OpenManage Enterprise
 version_added: "2.9.10"
 description: This module creates a baseline on OpenManage Enterprise.
 options:
@@ -83,6 +83,9 @@ options:
 requirements:
     - "python >= 2.7.5"
 author: "Jagadeesh N V(@jagadeeshnv)"
+notes:
+    - Run this module from a system that has direct access to DellEMC OpenManage Enterprise.
+    - This module does not support C(check_mode).
 '''
 
 EXAMPLES = r'''