+++ library/dellemc/ome/ome_network_vlan_info.py
@@ -3,15 +3,14 @@
 
 #
 # Dell EMC OpenManage Ansible Modules
-# Version 2.1.3
-# Copyright (C) 2020 Dell Inc. or its subsidiaries. All Rights Reserved.
+# Version 3.0.0
+# Copyright (C) 2020-2021 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 #
 
 
 from __future__ import (absolute_import, division, print_function)
-
 __metaclass__ = type
 
 ANSIBLE_METADATA = {'metadata_version': '1.1',
@@ -21,7 +20,7 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
 DOCUMENTATION = r'''
 ---
 module: ome_network_vlan_info
-short_description: Retrieves the information about networks VLAN(s) present in OpenManage Enterprise.
+short_description: Retrieves the information about networks VLAN(s) present in OpenManage Enterprise
 version_added: "2.9.14"
 description:
     This module allows to retrieve the following
@@ -58,6 +57,9 @@ options:
 requirements:
     - "python >= 2.7.5"
 author: "Deepak Joshi(@deepakjoshishri)"
+notes:
+    - Run this module from a system that has direct access to DellEMC OpenManage Enterprise.
+    - This module supports C(check_mode).
 '''
 
 EXAMPLES = """
@@ -235,7 +237,7 @@ def main():
             "name": {"required": False, "type": 'str'}
         },
         mutually_exclusive=[["id", "name"]],
-        supports_check_mode=False)
+        supports_check_mode=True)
     try:
         with RestOME(module.params, req_session=True) as rest_obj:
             # Form URI to fetch network VLAN information