+++ library/dellemc/ome/ome_smart_fabric.py
@@ -3,25 +3,20 @@
 
 #
 # Dell EMC OpenManage Ansible Modules
-# Version 2.1.5
-# Copyright (C) 2020 Dell Inc. or its subsidiaries. All Rights Reserved.
+# Version 3.0.0
+# Copyright (C) 2020-2021 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 #
 
 
 from __future__ import (absolute_import, division, print_function)
-
 __metaclass__ = type
-
-ANSIBLE_METADATA = {'metadata_version': '1.1',
-                    'status': ['preview'],
-                    'supported_by': 'community'}
 
 DOCUMENTATION = r'''
 ---
 module: ome_smart_fabric
-short_description: "Create, modify or delete a fabric on OpenManage Enterprise Modular."
+short_description: "Create, modify or delete a fabric on OpenManage Enterprise Modular"
 version_added: "2.9.14"
 description:
   - This module allows to create a fabric, and modify or delete an existing fabric
@@ -100,6 +95,9 @@ requirements:
     - "python >= 2.7.17"
 author:
     - "Sajna Shetty(@Sajna-Shetty)"
+notes:
+    - Run this module from a system that has direct access to DellEMC OpenManage Enterprise Modular.
+    - This module supports C(check_mode).
 '''
 
 EXAMPLES = r'''
@@ -227,6 +225,33 @@ SECONDARY_SWITCH_OVERLAP_MSG = "The switch details of the secondary switch overl
                                " switch details."
 
 
+def get_service_tag_with_fqdn(rest_obj, module):
+    """
+    get the service tag, if hostname is dnsname
+    """
+    hostname = module.params["hostname"]
+    service_tag = None
+    device_details = rest_obj.get_all_items_with_pagination(DEVICE_URI)
+    for each_device in device_details["value"]:
+        for item in each_device["DeviceManagement"]:
+            if item["DnsName"] == hostname:
+                return each_device["DeviceServiceTag"]
+    return service_tag
+
+
+def validate_lead_msm_version(each_domain, module, fabric_design=None):
+    """
+    validate lead chassis for design type
+    and find the msm version of the domain
+    """
+    role_type = each_domain["DomainRoleTypeValue"].upper()
+    if fabric_design and fabric_design == "2xMX9116n_Fabric_Switching_Engines_in_different_chassis" and \
+            role_type != "LEAD":
+        module.fail_json(msg=LEAD_CHASSIS_ERROR_MSG.format(fabric_design))
+    msm_version = each_domain["Version"]
+    return msm_version
+
+
 def get_msm_device_details(rest_obj, module):
     """
     Get msm details
@@ -239,15 +264,14 @@ def get_msm_device_details(rest_obj, module):
     hostname = module.params["hostname"]
     fabric_design = module.params.get("fabric_design")
     msm_version = ""
-    service_tag = None
+    service_tag = get_service_tag_with_fqdn(rest_obj, module)
     domain_details = rest_obj.get_all_items_with_pagination(DOMAIN_URI)
     for each_domain in domain_details["value"]:
+        if service_tag and service_tag == each_domain["Identifier"]:
+            msm_version = validate_lead_msm_version(each_domain, module, fabric_design)
+            break
         if hostname in each_domain["PublicAddress"]:
-            role_type = each_domain["DomainRoleTypeValue"].upper()
-            if fabric_design and fabric_design == "2xMX9116n_Fabric_Switching_Engines_in_different_chassis" and \
-                    role_type != "LEAD":
-                module.fail_json(msg=LEAD_CHASSIS_ERROR_MSG.format(fabric_design))
-            msm_version = each_domain["Version"]
+            msm_version = validate_lead_msm_version(each_domain, module, fabric_design)
             service_tag = each_domain["Identifier"]
             break
     else: