+++ library/dellemc/ome/ome_smart_fabric_uplink.py
@@ -3,8 +3,8 @@
 
 #
 # Dell EMC OpenManage Ansible Modules
-# Version 2.1.4
-# Copyright (C) 2020 Dell Inc. or its subsidiaries. All Rights Reserved.
+# Version 3.0.0
+# Copyright (C) 2020-2021 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 #
@@ -20,7 +20,7 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
 DOCUMENTATION = r'''
 ---
 module: ome_smart_fabric_uplink
-short_description: "Create, modify or delete a uplink for a fabric on OpenManage Enterprise Modular."
+short_description: Create, modify or delete a uplink for a fabric on OpenManage Enterprise Modular
 version_added: "2.9.15"
 description: This module allows to create, modify or delete an uplink for a fabric.
 options:
@@ -56,11 +56,11 @@ options:
     required: true
   name:
     type: str
-    description: Provide the I(name) of the uplink to be modified.
+    description: Provide the I(name) of the uplink to be created, modified or deleted.
     required: true
   new_name:
     type: str
-    description: Provide the I(name) of the uplink to be modified.
+    description: Provide the new I(new_name) for the uplink.
     required: false
   description:
     type: str
@@ -119,6 +119,9 @@ requirements:
     - "python >= 2.7.17"
 author:
     - "Jagadeesh N V(@jagadeeshnv)"
+notes:
+    - Run this module from a system that has direct access to DellEMC OpenManage Enterprise Modular.
+    - This module supports C(check_mode).
 '''
 
 EXAMPLES = r'''