+++ library/dellemc/ome/ome_template_identity_pool.py
@@ -3,8 +3,8 @@
 
 #
 # Dell EMC OpenManage Ansible Modules
-# Version 2.1.5
-# Copyright (C) 2020 Dell Inc. or its subsidiaries. All Rights Reserved.
+# Version 3.0.0
+# Copyright (C) 2020-2021 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 #
@@ -20,7 +20,7 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
 DOCUMENTATION = r'''
 ---
 module: ome_template_identity_pool
-short_description: Attach or detach an identity pool to a requested template on OpenManage Enterprise.
+short_description: Attach or detach an identity pool to a requested template on OpenManage Enterprise
 version_added: "2.9.10"
 description: This module allows to-
   - Attach an identity pool to a requested template on OpenManage Enterprise.
@@ -54,6 +54,9 @@ options:
 requirements:
     - "python >= 2.7.5"
 author: "Felix Stephen (@felixs88)"
+notes:
+    - Run this module from a system that has direct access to DellEMC OpenManage Enterprise.
+    - This module does not support C(check_mode).
 '''
 
 EXAMPLES = r'''
@@ -173,6 +176,7 @@ def main():
             "template_name": {"required": True, "type": "str"},
             "identity_pool_name": {"required": False, "type": "str"},
         },
+        supports_check_mode=False
     )
     try:
         with RestOME(module.params, req_session=True) as rest_obj: